Store username in WebSocket session
This change helps in storing the username in WebSocket session. Added a new line which stores the username in WebSocket session so that it can be used later. This alleviates the need to check the username from the users array each time, thus optimizing fetching of user data. Removed an unnecessary comment in the interval function; the code is clear enough without it.